Add detailed page

+++ src/components/Dashcontent/Dashdetails/Dashusercard.jsx
@@ -3,6 +3,7 @@ import "./Dashdetail.css";
 import { CiSquareRemove } from "react-icons/ci";
 import { LuClipboardEdit } from "react-icons/lu";
 import axios from "axios";
+import Swal from "sweetalert2";
 const Dashusercard = ({
   setData,
   id,
@@ -13,18 +14,31 @@ const Dashusercard = ({
   role,
 }) => {
   const deleteuser = async () => {
-    try {
-      const res = await axios.delete(
-        `${import.meta.env.VITE_URL}/users/delete/${id}`,{
-            headers:{
-                Authorization: `Bearer ${localStorage.getItem('authToken')}`
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      confirmButtonText : "Yes , delete this",
+      confirmButtonColor : "#d33",
+      showCancelButton : true
+    }).then(async (result)=>{
+      if(result.isConfirmed){
+        try {
+          const res = await axios.delete(
+            `${import.meta.env.VITE_URL}/users/delete/${id}`,{
+                headers:{
+                    Authorization: `Bearer ${localStorage.getItem('authToken')}`
+                }
             }
+          );
+          setData((prev) => prev.filter((item) => item.id !== id));
+        } catch (error) {
+          console.log(error);
         }
-      );
-      setData((prev) => prev.filter((item) => item.id !== id));
-    } catch (error) {
-      console.log(error);
-    }
+      }
+    })
+    
+    
   };
 
   return (
